@@ -1,75 +1,74 @@
 {
-    "name": "bioengine-admin",
-    "version": "3.0.0",
-    "license": "MIT",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/BioWareRu/BioEngine.Admin.git"
-    },
-    "bugs": {
-        "url": "https://github.com/BioWareRu/BioEngine.Admin/issues"
-    },
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "start-hmr": "ng serve --configuration hmr --source-map=false --hmr-warning=false",
-        "start-hmr-sourcemaps": "ng serve --configuration hmr --source-map=true --hmr-warning=false",
-        "build": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build",
-        "build-stats": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --stats-json",
-        "build-prod": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod",
-        "build-prod-stats": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod --stats-json",
-        "lint": "ng lint",
-        "bundle-report": "webpack-bundle-analyzer dist/stats.json"
-    },
-    "dependencies": {
-        "@angular/animations": "8.0.3",
-        "@angular/cdk": "8.0.1",
-        "@angular/common": "8.0.3",
-        "@angular/compiler": "8.0.3",
-        "@angular/core": "8.0.3",
-        "@angular/flex-layout": "8.0.0-beta.26",
-        "@angular/forms": "8.0.3",
-        "@angular/material": "8.0.1",
-        "@angular/material-moment-adapter": "8.0.1",
-        "@angular/platform-browser": "8.0.3",
-        "@angular/platform-browser-dynamic": "8.0.3",
-        "@ckeditor/ckeditor5-angular": "1.1.0",
-        "@ckeditor/ckeditor5-build-balloon": "12.2.0",
-        "@fortawesome/fontawesome-free": "5.9.0",
-        "moment": "2.24.0",
-        "ngx-custom-validators": "8.0.0",
-        "ngx-input-file": "8.0.0",
-        "ngx-moment": "3.4.0",
-        "object-path": "0.11.4",
-        "rxjs": "6.5.2",
-        "tslib": "1.10.0",
-        "url-parse": "1.4.7",
-        "uuid": "3.3.2",
-        "zone.js": "0.9.1",
-        "@angular/router": "8.0.3",
-        "angular-oauth2-oidc": "5.0.2",
-        "angular-tree-component": "8.4.0",
-        "bioengine-angular": "1.0.17",
-        "classlist.js": "1.1.20150312",
-        "perfect-scrollbar": "1.4.0",
-        "web-animations-js": "github:angular/web-animations-js#release_pr208"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "0.800.3",
-        "@angular/cli": "8.0.4",
-        "@angular/compiler-cli": "8.0.3",
-        "@angular/language-service": "8.0.3",
-        "@types/node": "12.0.10",
-        "@types/object-path": "0.11.0",
-        "@types/twitter-for-web": "0.0.1",
-        "@types/url-parse": "1.4.3",
-        "@types/uuid": "3.4.4",
-        "codelyzer": "5.1.0",
-        "node-sass": "4.12.0",
-        "ts-node": "8.3.0",
-        "tslint": "5.18.0",
-        "tslint-consistent-codestyle": "1.15.1",
-        "typescript": "3.4.5",
-        "webpack-bundle-analyzer": "3.3.2"
-    }
+  "name": "bioengine-admin",
+  "version": "3.0.0",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/BioWareRu/BioEngine.Admin.git"
+  },
+  "bugs": {
+    "url": "https://github.com/BioWareRu/BioEngine.Admin/issues"
+  },
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "start-hmr": "ng serve --configuration hmr --source-map=false --hmr-warning=false",
+    "start-hmr-sourcemaps": "ng serve --configuration hmr --source-map=true --hmr-warning=false",
+    "build": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build",
+    "build-stats": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --stats-json",
+    "build-prod": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod",
+    "build-prod-stats": "node --max_old_space_size=6144 ./node_modules/@angular/cli/bin/ng build --prod --stats-json",
+    "lint": "ng lint",
+    "bundle-report": "webpack-bundle-analyzer dist/stats.json"
+  },
+  "dependencies": {
+    "@angular/animations": "8.2.0",
+    "@angular/cdk": "8.1.2",
+    "@angular/common": "8.2.0",
+    "@angular/compiler": "8.2.0",
+    "@angular/core": "8.2.0",
+    "@angular/flex-layout": "8.0.0-beta.26",
+    "@angular/forms": "8.2.0",
+    "@angular/material": "8.1.2",
+    "@angular/material-moment-adapter": "8.1.2",
+    "@angular/platform-browser": "8.2.0",
+    "@angular/platform-browser-dynamic": "8.2.0",
+    "@ckeditor/ckeditor5-angular": "1.1.0",
+    "@ckeditor/ckeditor5-build-balloon": "12.3.1",
+    "@fortawesome/fontawesome-free": "5.10.1",
+    "moment": "2.24.0",
+    "ngx-custom-validators": "8.0.0",
+    "ngx-input-file": "8.0.0",
+    "ngx-moment": "3.4.0",
+    "object-path": "0.11.4",
+    "rxjs": "6.5.2",
+    "tslib": "1.10.0",
+    "url-parse": "1.4.7",
+    "uuid": "3.3.2",
+    "zone.js": "0.9.1",
+    "@angular/router": "8.2.0",
+    "angular-oauth2-oidc": "8.0.4",
+    "bioengine-angular": "1.2.0",
+    "classlist.js": "1.1.20150312",
+    "perfect-scrollbar": "1.4.0",
+    "web-animations-js": "github:angular/web-animations-js#release_pr208"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "0.802.0",
+    "@angular/cli": "8.2.0",
+    "@angular/compiler-cli": "8.2.0",
+    "@angular/language-service": "8.2.0",
+    "@types/node": "12.6.9",
+    "@types/object-path": "0.11.0",
+    "@types/twitter-for-web": "0.0.1",
+    "@types/url-parse": "1.4.3",
+    "@types/uuid": "3.4.5",
+    "codelyzer": "5.1.0",
+    "node-sass": "4.12.0",
+    "ts-node": "8.3.0",
+    "tslint": "5.18.0",
+    "tslint-consistent-codestyle": "1.15.1",
+    "typescript": "3.4.5",
+    "webpack-bundle-analyzer": "3.4.1"
+  }
 }
