@@ -1,5 +1,7 @@
 import { NgModule } from '@angular/core';
-import { MatCardModule, MatGridListModule } from '@angular/material';
+import { MatListModule } from '@angular/material/list';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatCardModule } from '@angular/material/card';
 
 import { RouterModule, Routes } from '@angular/router';
 import { BrcCommonModule } from '@common/BRCCommonModule';
@@ -29,6 +31,7 @@ const routes: Routes = [{
         RouterModule.forChild(routes),
         MatGridListModule,
         MatCardModule,
+        MatListModule
     ],
     exports: [
         RouterModule
