@@ -1,14 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { AbstractListComponent } from '@common/list/AbstractListComponent';
 import { PageContext } from '@common/PageContext';
-import { Icon, ListTableColumn, ListTableColumnAction, ListTableColumnType, Menu, MenuService, SiteTableColumn } from 'bioengine-angular';
+import { Icon, ListTableColumn, ListTableColumnAction, ListTableColumnType, Menu, MenuService, SitesTableColumn } from 'bioengine-angular';
 
 @Component({
     selector: 'menu-list',
     templateUrl: './MenuListComponent.html',
     providers: [PageContext]
 })
-export class MenuListComponent extends AbstractListComponent<Menu> implements OnInit {
+export class MenuListComponent extends AbstractListComponent<Menu> {
     constructor(context: PageContext, menuService: MenuService) {
         super(menuService, context);
 
@@ -26,7 +26,7 @@ export class MenuListComponent extends AbstractListComponent<Menu> implements On
                 'Дата',
                 ListTableColumnType.TimeAgo
             ).setSortable(),
-            new SiteTableColumn<Menu>('siteId', 'Сайт'),
+            new SitesTableColumn<Menu>('siteIds', 'Сайт'),
             new ListTableColumn<Menu>('actions', '').addAction(
                 new ListTableColumnAction<Menu>('Удалить', new Icon('fa-trash')).setClick(menu =>
                     this.deleteItem(menu)
