@@ -1,15 +1,19 @@
+import { DragDropModule } from '@angular/cdk/drag-drop';
 import { NgModule } from '@angular/core';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatTreeModule } from '@angular/material/tree';
 
 import { RouterModule, Routes } from '@angular/router';
 import { BrcFormsModule } from '@common/forms/BrcFormsModule';
 import { BrcListModule } from '@common/list/BrcListModule';
-import { TreeModule } from 'angular-tree-component';
 import { BioCommonModule, BioDialogsModule, DialogService } from 'bioengine-angular';
 import { MenuFormPageComponent } from './MenuFormPageComponent';
 import { MenuFormComponent } from './MenuFormComponent';
 import { MenuListComponent } from './MenuListComponent';
 import { MenuItemFormComponent } from './MenuItemFormComponent';
-import { MenuItemFormDialogComponent } from './MenuItemFormDialogComponent';
 
 const routes: Routes = [
     {
@@ -38,16 +42,19 @@ const routes: Routes = [
         BrcFormsModule,
         BrcListModule,
         RouterModule.forChild(routes),
-        TreeModule.forRoot()
+        MatTreeModule,
+        MatCheckboxModule,
+        MatIconModule,
+        MatFormFieldModule,
+        MatInputModule,
+        DragDropModule
     ],
     declarations: [
         MenuListComponent,
         MenuFormComponent,
         MenuItemFormComponent,
-        MenuItemFormDialogComponent,
         MenuFormPageComponent
     ],
-    entryComponents: [MenuItemFormDialogComponent],
     providers: [DialogService]
 })
 export class MenuModule {
